refactor(portal-base): move promptForDeployPath to its own file
So that it can be reused.
