+++ projects/js-toolkit/packages/portal-base/src/deploy.ts
@@ -3,11 +3,12 @@
  * SPDX-License-Identifier: LGPL-3.0-or-later
  */
 
-import {FilePath, Project, format} from '@liferay/js-toolkit-core';
+import {Project, format} from '@liferay/js-toolkit-core';
 import fs from 'fs';
-import {createInterface} from 'readline';
 
-const {fail, print, question, success} = format;
+import promptForDeployPath from './util/promptForDeployPath';
+
+const {fail, print, success} = format;
 
 export default async function deploy(): Promise<void> {
 	const project = new Project('.');
@@ -40,32 +41,4 @@ export default async function deploy(): Promise<void> {
 			project.dist.file.basename().asNative
 		}} deployed to {${deployDir.asNative}}`
 	);
-}
-
-async function promptForDeployPath(project: Project): Promise<FilePath> {
-	const lines = createInterface({
-		input: process.stdin,
-	});
-
-	print(question`Please enter your local Liferay installation directory`);
-
-	let deployDir: FilePath;
-
-	for await (const line of lines) {
-		deployDir = new FilePath(line).join('osgi', 'modules').resolve();
-
-		if (fs.existsSync(deployDir.asNative)) {
-			project.deploy.storeDir(deployDir);
-
-			break;
-		}
-		else {
-			print(fail`${deployDir.asNative} does not exist`);
-			print(
-				question`Please enter your local Liferay installation directory`
-			);
-		}
-	}
-
-	return deployDir;
 }
